+++ packages/console-sankey-chart/src/index.ts
@@ -46,6 +46,7 @@ const cfg = {
       nodeAlign = 'justify',
       nodeWidth = 0.02,
       nodePadding = 0.02,
+      topology = false,
     } = config;
 
     const d3SankeyConfig = {
@@ -54,11 +55,22 @@ const cfg = {
       nodePadding,
     };
 
-    transform(adaptData, d3SankeyConfig);
+    const customOptions = {
+      topology,
+    };
+
+    transform(adaptData, d3SankeyConfig, customOptions);
 
     g2Legend(chart, config, config.legend);
 
-    g2Tooltip(chart, config, config.tooltip);
+    let tooltip = {
+      showTitle: false,
+    };
+    if (config.tooltip) {
+      tooltip = config.tooltip;
+    }
+
+    g2Tooltip(chart, config, tooltip);
 
     chart.axis(false);
     chart.scale({
@@ -77,7 +89,19 @@ const cfg = {
       .position(position)
       .shape(linkType)
       .color(linkColor)
-      .opacity(linkOpacity);
+      .opacity(linkOpacity)
+      .tooltip('source*target', (source, target) => {
+        let name = source.index;
+        let value = target.index;
+        if (source.name !== undefined && target.name !== undefined) {
+          name = source.name;
+          value = target.name;
+        }
+        return {
+          name,
+          value,
+        };
+      })
 
     g2Size(edgeView, config, config.size);
     g2Style(edgeView, config, config.style);
@@ -94,11 +118,20 @@ const cfg = {
       .color('name', colors)
       .style({
         stroke: '#ccc'
-      });
+      })
+      // .label('name', {
+      //   useHtml: true,
+      //   htmlTemplate: (text, item, index) => {
+      //     return `<div>${text}</div>`;
+      //   },
+      // })
+      .tooltip('name', name => ({
+        name,
+      }))
 
     g2Size(nodeView, config, config.size);
     g2Style(nodeView, config, config.style);
-    g2Label(nodeView, config, config.label);
+    // g2Label(nodeView, config, config.label);
 
     chart.render();
   },